[#1]: Add usage examples to documenetation
