--- docs/cli_subcommand.rst
@@ -1,7 +1,0 @@
-Fellowship command
-***********************
-
-.. argparse::
-   :module: fellowship.cli
-   :func: get_parser
-   :prog: fellowship
