+++ docs/spelling_wordlist.txt
@@ -3,19 +3,25 @@ argparser
 cli
 config
 configs
+const
 consts
+dir
 enums
+env
 gRPC
 github
 jinja
 json
 kwargs
 params
+pre
 readme
+repo
 rtype
 submodules
 tox
 url
 utilis
+validators
 whitespace
 yaml
