+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =py39,
          pylint39,
-         pip-compile
+         pip-compile,
          docs-{dirhtml,spelling}
